Added feature to the system tray context menu for custom actions for retraining the model.
Added feature to the system tray context menu for custom actions. Once you updated the custom actions function, training data, or enabled/disabled actions you can now re-train the model from this menu as well. It will close Mykros and train the model, once it is complete it will automatically re-launch Mykros.

Fixed a bug that caused the enabled(true)/disabled(false) translation setting in the settings.yml file to default to true if an invalid value was detected instead of defaulting to false.